requirements.txt
streamlit
google-generativeai
pandas
gspread
oauth2client
pypdf
python-docx
google-api-python-client
